requirements.txt.txt
pytest==8.3.5
pyinstaller==6.6.0
pycryptodome==3.20.0
flet==0.21.0